test(task-migrate-clean-batch): 迁 defaults_panel 的 2 个非控件用例到 gui_unit_test
420.5 Step 8。部分迁移 2/35，其余 33 个驱动真实 Settings modal 控件，接缝在
UI 侧，留在 gui_test。

- inline_ac3_registry_covers_every_row —— 字段编辑器注册表与出厂文档真实行集
  的双向一致（每行要么注册且 kind 符合预期、要么显式不注册；且不存在文档
  产不出的注册项）
- inline_ac5b_row_tint_predicate_four_states —— RowWouldChangeOnSave 的四态

两处必要改动，断言零改动：
- ac3 的 ResetTestState()：真正让 MakeNewDocumentState() 确定的是紧接着那个
  新建空目录的 ScopedUserConfigSource，ResetTestState 是它周围的 harness 脚手架
- ac3 改用局部 GuiState：BuildDefaultDiffRows 本就把要读的 state 当参数收，
  绕全局只是多给邻居一条改写本用例答案的路（plan §3「参数化纯函数改局部状态」）

核验：源文件侧断言减量 IM_CHECK 6 + IM_CHECK_SILENT 5 + IM_CHECK_EQ 1 = 12，
新文件 EXPECT_TRUE 11 + EXPECT_EQ 1 = 12，按宏形态逐类对上；
gui_test 静态用例数 506→504；gui_unit_test 69→71 全绿。
